Fix time range of update_resources stored proc
The filter for events to be included was excluding all fields as
requiring events start time to be before _from_date.
Limits have been adjusted too be after _from_date and before _run_date
(NOW). This may need revisiting as I am not certain of the original
intent of the filter.